combined modal with createevent component
