+++ src/app/page.tsx
@@ -3,7 +3,6 @@ import React, { useState } from 'react';
 import Navbar from "@components/Navbar";
 import Calendar from "@components/Calendar";
 import CreateEvent from '@components/CreateEvent/CreateEvent';
-import CreateEventButton from '@components/CreateEvent/CreateEventButton';
 
 export default function Home() {
   const [admin, setAdmin] = useState(false);
@@ -20,7 +19,7 @@ export default function Home() {
       >
         <Navbar />
         <div style={{ width: "70%", margin: "20px" }}>
-          <CreateEventButton/>
+          <CreateEvent/>
           <Calendar admin={admin} />
         </div>
       </div>
